+++ tests/testapp/admin.py
@@ -1,7 +1,29 @@
+from django import forms
 from django.contrib import admin
 
+from django_prose_editor.widgets import ProseEditorWidget
 from testapp import models
 
 
-admin.register(models.ProseEditorModel)
-admin.register(models.SanitizedProseEditorModel)
+@admin.register(models.ProseEditorModel)
+class ProseEditorModelAdmin(admin.ModelAdmin):
+    pass
+
+
+class SanitizedProseEditorModelForm(forms.ModelForm):
+    class Meta:
+        widgets = {
+            "description": ProseEditorWidget(
+                config={
+                    "types": ["paragraph", "strong", "em", "link", "heading"],
+                    "history": False,
+                    "html": False,
+                    "typographic": True,
+                }
+            )
+        }
+
+
+@admin.register(models.SanitizedProseEditorModel)
+class SanitizedProseEditorModelAdmin(admin.ModelAdmin):
+    form = SanitizedProseEditorModelForm